Refactor scripts.rs to split out the callbacks mod into its own file and to separate the Message handlers into their own fns. - Fix fn for Message::SetParameter so that it really does restart the script if the on_set_parameter Lua function isn't present.  Also, save the parameter changes first, in case the script chokes on the parameter update. - Fix documentation on mouse button function names. - Simplify how parameters are passed from main.rs to threads.rs to script.rs.  More changes to come later. - Tweak the way Lua functions are called so that the functions are now cached.  Caching probably isn't necessary for performance, but was just a side effect of knowing when to skip the function call if the function was absent. - Also, remove ACTIVE_SCRIPTS.  It was not used except where elements were being added to it and never removed from it.  (If the need for it returns, it should come back as something like a map of structs that also contains the Message channel sender.) - Finally, tweak the Makefile to avoid using brace expansion.  That only works in some shells.
