+++ docs/LIBRARY.md
@@ -219,8 +219,8 @@ Eruption currently calls the following event handler functions, if they are pres
 | `on_apply_parameter(parameter, value)` | _core_     | parameter: The name of the parameter, value: The new value of the parameter                                                                                                                                    | Sent whenever a script parameter shall be updated |
 | `on_key_down(key_index)`               | _Keyboard_ | key_index: Key index (column major order)                                                                                                                                                                      |                                                   |
 | `on_key_up(key_index)`                 | _Keyboard_ | key_index: Key index (column major order)                                                                                                                                                                      |                                                   |
-| `on_mouse_down(button_index)`          | _Mouse_    | button_index: Index of mouse button                                                                                                                                                                            |                                                   |
-| `on_mouse_up(button_index)`            | _Mouse_    | button_index: Index of mouse button                                                                                                                                                                            |                                                   |
+| `on_mouse_button_down(button_index)`   | _Mouse_    | button_index: Index of mouse button                                                                                                                                                                            |                                                   |
+| `on_mouse_button_up(button_index)`     | _Mouse_    | button_index: Index of mouse button                                                                                                                                                                            |                                                   |
 | `on_mouse_wheel(direction)`            | _Mouse_    | direction: 1 == up, 2 == down, 3 == left, 4 == right                                                                                                                                                           |                                                   |
 | `on_mouse_move(rel_x, rel_y, rel_z)`   | _Mouse_    | x, y, z coordinate updates                                                                                                                                                                                     | Coordinates are relative (delta values)           |
 | `on_hid_event(event_type, arg1)`       | _Hardware_ | event_type: 0 == unknown, 1 == KeyUp, 2 == KeyDown, 3 == MuteButton, 4 == Volume knob, 5 == Brightness knob, 6 == Set Brightness, 7 == Next/previous Profile, arg1: data payload e.g.: scan codes/status codes |                                                   |
